Rename auth.js to api.js
