+++ frontend/src/components/WishStep.js
@@ -1,6 +1,6 @@
 import React, { useState, useContext } from "react";
 import { FileEntry, FileCategory, FileListing, FileButton } from "./FileEntry";
-import { api_category_download, api_file_display } from "../auth";
+import { api_category_download, api_file_display } from "../api";
 import ModalVariants from "./ModalVariants";
 import "../styles/WishGrantingPage.css";
 import FileStructure from "./FileStructure";
